Add files via upload
Task 1: Mobile App Signup Flow 
A mobile app signup flow is the process of users creating an account and signing up for an
app. The goal of a good signup flow is to make it easy and quick for users to sign up, while
also collecting the necessary information from them.
A good signup flow should be:
Short and sweet: Users are more likely to complete a signup flow if it is short and easy to
follow.
Personalized: The more personalized the signup flow is, the more likely users are to
complete it. You can personalize the experience by using the user
'
s name, location, or

interests.
Useful: The signup flow should collect the necessary information from users, but it should
not ask for too much information.
Error-free: The signup flow should be free of errors, so that users do not have to repeat the
process.
Secure: The signup flow should be secure, so that users
'
personal information is protected.